LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.BlobURLDedicatedWorker.sub.https.tentative.window-expected.txt


FAIL Verify that if the third-party context creates a blob URL using the StorageAccessHandle and sends it to the dedicated worker, the dedicated worker fetch succeeds. assert_equals: expected "Blob URL DedicatedWorker tests completed successfully." but got "Unable to load handle in same-origin context for BlobURLDedicatedWorker"
